Update progress counters

+++ app/src/main/java/com/omar/retromp3recorder/app/ui/files/preview/CurrentFileView.kt
@@ -15,7 +15,7 @@ object CurrentFileView {
     )
 
     sealed class Input {
-        data class SeekToPosition(val position: Int) : Input()
+        data class SeekToPosition(val position: Long) : Input()
         object SeekingStarted : Input()
         object SeekingFinished : Input()
     }
